+++ seanime-web/src/app/(main)/server-data-wrapper.tsx
@@ -9,6 +9,7 @@ import { Card } from "@/components/ui/card"
 import { defineSchema, Field, Form } from "@/components/ui/form"
 import { logger } from "@/lib/helpers/debug"
 import { ANILIST_OAUTH_URL, ANILIST_PIN_URL } from "@/lib/server/config"
+import Link from "next/link"
 import { usePathname, useRouter } from "next/navigation"
 import React from "react"
 
@@ -118,21 +119,26 @@ export function ServerDataWrapper(props: ServerDataWrapperProps) {
                             <img src="/logo.png" alt="logo" className="w-24 h-auto" />
                         </div>
                         <h3>Welcome!</h3>
-                        <Button
-                            onClick={() => {
-                                window.open(ANILIST_PIN_URL, "_blank")
-                            }}
-                            leftIcon={<svg
-                                xmlns="http://www.w3.org/2000/svg" fill="currentColor" width="24" height="24"
-                                viewBox="0 0 24 24" role="img"
-                            >
-                                <path
-                                    d="M6.361 2.943 0 21.056h4.942l1.077-3.133H11.4l1.052 3.133H22.9c.71 0 1.1-.392 1.1-1.101V17.53c0-.71-.39-1.101-1.1-1.101h-6.483V4.045c0-.71-.392-1.102-1.101-1.102h-2.422c-.71 0-1.101.392-1.101 1.102v1.064l-.758-2.166zm2.324 5.948 1.688 5.018H7.144z"
-                                />
-                            </svg>}
-                            intent="white"
-                            size="md"
-                        >Get AniList token</Button>
+                        <Link
+                            href={ANILIST_PIN_URL}
+                            target="_blank"
+                        >
+                            <Button
+                                // onClick={() => {
+                                //     window.open(ANILIST_PIN_URL, "_blank")
+                                // }}
+                                leftIcon={<svg
+                                    xmlns="http://www.w3.org/2000/svg" fill="currentColor" width="24" height="24"
+                                    viewBox="0 0 24 24" role="img"
+                                >
+                                    <path
+                                        d="M6.361 2.943 0 21.056h4.942l1.077-3.133H11.4l1.052 3.133H22.9c.71 0 1.1-.392 1.1-1.101V17.53c0-.71-.39-1.101-1.1-1.101h-6.483V4.045c0-.71-.392-1.102-1.101-1.102h-2.422c-.71 0-1.101.392-1.101 1.102v1.064l-.758-2.166zm2.324 5.948 1.688 5.018H7.144z"
+                                    />
+                                </svg>}
+                                intent="white"
+                                size="md"
+                            >Get AniList token</Button>
+                        </Link>
 
                         <Form
                             schema={defineSchema(({ z }) => z.object({